refactor: modularize send and receive logic into dedicated files
